Created a generalised version of an error file
All the file errors seem to share the same shape from the responses I've seen. Therefore, as we add more error types it's probably better to hoist this out rather than repeat the same code
